+++ mne_bids/path.py
@@ -10,16 +10,145 @@ from collections import OrderedDict
 from copy import deepcopy
 from os import path as op
 from pathlib import Path
+from datetime import datetime
 
+import numpy as np
 from mne.utils import warn, logger
 
 from mne_bids.config import (
     ALLOWED_PATH_ENTITIES, ALLOWED_FILENAME_EXTENSIONS,
     ALLOWED_FILENAME_SUFFIX, ALLOWED_PATH_ENTITIES_SHORT,
-    ALLOWED_DATATYPES, SUFFIX_TO_DATATYPE, ALLOWED_DATATYPE_EXTENSIONS)
+    ALLOWED_DATATYPES, SUFFIX_TO_DATATYPE, ALLOWED_DATATYPE_EXTENSIONS,
+    reader)
 from mne_bids.utils import (_check_key_val, _check_empty_room_basename,
                             _check_types, param_regex,
                             _ensure_tuple)
+
+
+def _get_matched_empty_room(bids_path):
+    """Get matching empty-room file for an MEG recording."""
+    if not isinstance(bids_path, BIDSPath):
+        raise RuntimeError('"bids_path" must be a BIDSPath object. Please '
+                           'instantiate using mne_bids.BIDSPath().')
+
+    # Check whether we have a BIDS root.
+    bids_root = bids_path.root
+    if bids_root is None:
+        raise ValueError('The root of the "bids_path" must be set. '
+                         'Please use `bids_path.update(root="<root>")` '
+                         'to set the root of the BIDS folder to read.')
+
+    from mne_bids import read_raw_bids  # avoid circular import.
+    bids_path = bids_path.copy()
+
+    datatype = 'meg'  # We're only concerned about MEG data here
+    bids_fname = bids_path.update(suffix=datatype,
+                                  root=bids_root).fpath
+    _, ext = _parse_ext(bids_fname)
+    if ext == '.fif':
+        extra_params = dict(allow_maxshield=True)
+    else:
+        extra_params = None
+
+    raw = read_raw_bids(bids_path=bids_path, extra_params=extra_params)
+    if raw.info['meas_date'] is None:
+        raise ValueError('The provided recording does not have a measurement '
+                         'date set. Cannot get matching empty-room file.')
+
+    ref_date = raw.info['meas_date']
+    if not isinstance(ref_date, datetime):
+        # for MNE < v0.20
+        ref_date = datetime.fromtimestamp(raw.info['meas_date'][0])
+
+    emptyroom_dir = Path(make_bids_folders(bids_root=bids_root,
+                                           subject='emptyroom',
+                                           make_dir=False))
+
+    if not emptyroom_dir.exists():
+        return None
+
+    # Find the empty-room recording sessions.
+    emptyroom_session_dirs = [x for x in emptyroom_dir.iterdir()
+                              if x.is_dir() and str(x.name).startswith('ses-')]
+    if not emptyroom_session_dirs:  # No session sub-directories found
+        emptyroom_session_dirs = [emptyroom_dir]
+
+    # Now try to discover all recordings inside the session directories.
+
+    allowed_extensions = list(reader.keys())
+    # `.pdf` is just a "virtual" extension for BTi data (which is stored inside
+    # a dedicated directory that doesn't have an extension)
+    del allowed_extensions[allowed_extensions.index('.pdf')]
+
+    candidate_er_fnames = []
+    for session_dir in emptyroom_session_dirs:
+        dir_contents = glob.glob(op.join(session_dir, datatype,
+                                         f'sub-emptyroom_*_{datatype}*'))
+        for item in dir_contents:
+            item = Path(item)
+            if ((item.suffix in allowed_extensions) or
+                    (not item.suffix and item.is_dir())):  # Hopefully BTi?
+                candidate_er_fnames.append(item.name)
+
+    # Walk through recordings, trying to extract the recording date:
+    # First, from the filename; and if that fails, from `info['meas_date']`.
+    best_er_bids_path = None
+    min_delta_t = np.inf
+    date_tie = False
+
+    failed_to_get_er_date_count = 0
+    for er_fname in candidate_er_fnames:
+        params = get_entities_from_fname(er_fname)
+        er_meas_date = None
+        params.pop('subject')  # er subject entity is different
+        er_bids_path = BIDSPath(subject='emptyroom', **params, datatype='meg',
+                                root=bids_root, check=False)
+
+        # Try to extract date from filename.
+        if params['session'] is not None:
+            try:
+                er_meas_date = datetime.strptime(params['session'], '%Y%m%d')
+            except (ValueError, TypeError):
+                # There is a session in the filename, but it doesn't encode a
+                # valid date.
+                pass
+
+        if er_meas_date is None:  # No luck so far! Check info['meas_date']
+            _, ext = _parse_ext(er_fname)
+            if ext == '.fif':
+                extra_params = dict(allow_maxshield=True)
+            else:
+                extra_params = None
+
+            er_raw = read_raw_bids(bids_path=er_bids_path,
+                                   extra_params=extra_params)
+
+            er_meas_date = er_raw.info['meas_date']
+            if er_meas_date is None:  # There's nothing we can do.
+                failed_to_get_er_date_count += 1
+                continue
+
+        er_meas_date = er_meas_date.replace(tzinfo=ref_date.tzinfo)
+        delta_t = er_meas_date - ref_date
+
+        if abs(delta_t.total_seconds()) == min_delta_t:
+            date_tie = True
+        elif abs(delta_t.total_seconds()) < min_delta_t:
+            min_delta_t = abs(delta_t.total_seconds())
+            best_er_bids_path = er_bids_path
+            date_tie = False
+
+    if failed_to_get_er_date_count > 0:
+        msg = (f'Could not retrieve the empty-room measurement date from '
+               f'a total of {failed_to_get_er_date_count} recording(s).')
+        warn(msg)
+
+    if date_tie:
+        msg = ('Found more than one matching empty-room measurement with the '
+               'same recording date. Selecting the first match.')
+        warn(msg)
+
+    return best_er_bids_path
 
 
 class BIDSPath(object):
@@ -487,6 +616,21 @@ class BIDSPath(object):
                 raise ValueError(f'Suffix {suffix} is not allowed. '
                                  f'Use one of these suffixes '
                                  f'{ALLOWED_FILENAME_SUFFIX}.')
+
+    def find_empty_room(self):
+        """Find the corresponding empty-room file of an MEG recording.
+
+        This will only work if the ``.root`` attribute of the
+        :class:`mne_bids.BIDSPath` instance has been set.
+
+        Returns
+        -------
+        BIDSPath | None
+            The path corresponding to the best-matching empty-room measurement.
+            Returns None if none was found.
+
+        """
+        return _get_matched_empty_room(self)
 
 
 def _get_matching_bidspaths_from_filesystem(bids_path):